feat: add constants module
